+++ src/test/kotlin/no/nav/tilleggsstonader/sak/brev/brevmottaker/MottakerTestUtil.kt
@@ -4,7 +4,7 @@ import no.nav.tilleggsstonader.sak.brev.brevmottaker.domain.Mottaker
 import no.nav.tilleggsstonader.sak.brev.brevmottaker.domain.MottakerRolle
 import no.nav.tilleggsstonader.sak.brev.brevmottaker.domain.MottakerType
 
-object MottakerUtil {
+object MottakerTestUtil {
 
     fun mottakerPerson(ident: String, mottakerRolle: MottakerRolle = MottakerRolle.BRUKER) = Mottaker(
         ident = ident,
